Add TomTom, with basic examples
Includes an intro section in the report,
and examples for basic create map
and setting map location.
Note: TomTom CSS is commented out because it duplicates leaflet/mapbox.
Need to investigate further what the differences are.

Also adds it to the example template,
and to example pages made from that template.

Related to #49